Relocate the sync engine into domains/sync (whole-file, re-export facade)
googleSync.ts — the crown-jewels sync loop (generation-sweep, lease
mutual-exclusion, adaptive cadence, snapshot replacement, ~1.8k lines) — moved
byte-for-byte via git mv into domains/sync/engine.ts, with a root googleSync.ts
that re-exports it. Chosen over a hand-split maximal reorg precisely because a
transcription slip in this file would be a SILENT sync corruption; git mv
carries zero such risk.

Verified the re-export preserves registration: `convex dev --once` deploys
cleanly (Convex validates cron references, so internal.googleSync.enqueueSyncs
et al. must resolve), types pass, and the moved sync itests (lease/generation/
cadence/sweep, 7) pass through internal.googleSync.*. api.d.ts diff is purely
additive; every api.googleSync.* / internal.googleSync.* path — and the helpers
calendarOps/calendarSync import — is unchanged. The itest's import.meta.glob was
repointed to the convex root so convex-test still loads the full surface.

The fine-grained split of engine.ts (orchestrator/eventSync/contactsSync/…) can
follow later with no path risk, since it's all under domains/sync now.
